Create a auth user service hook
