+++ packages/client-web/src/modules/auth-user/model/index.ts
@@ -1,11 +1,10 @@
 import {
-  AuthUserService,
   AuthUserState,
   useAuthUserModel as _useAuthUserModel,
   makeReduxStoreSlice,
 } from "@quoll/client-lib";
-import { useMemo } from "react";
 
+import { useAuthUserService } from "../service";
 import { RootState } from "store";
 
 const defaultState: AuthUserState = {
@@ -19,13 +18,7 @@ export const authUserStore = makeReduxStoreSlice<AuthUserState, RootState>(
 );
 
 export const useAuthUserModel = (getAccessToken: () => Promise<string>) => {
-  const service = useMemo(() => {
-    if (process.env.REACT_APP_API_URL === undefined) {
-      throw new Error("REACT_APP_API_URL is not defined");
-    }
-
-    return new AuthUserService(getAccessToken, process.env.REACT_APP_API_URL);
-  }, [getAccessToken]);
+  const service = useAuthUserService(getAccessToken);
 
   return _useAuthUserModel(authUserStore.useStore, service);
 };
